A simple game of guess the number
The game has an easy mode and a hard mode. The user is able to enter as
many guesses as they want in easy mode, and only three tries in the hard
mode. The program asks at the beginning before each play if you'd like
to be in easy or hard mode. Uses the random libraryl, loops, and
conditional statements.